+++ CHANGELOG.md
@@ -2,7 +2,6 @@
 
 # [0.2.0](https://github.com/JuliaGNSS/RINEXParser.jl/compare/v0.1.0...v0.2.0) (2026-08-03)
 
-
 ### Features
 
-* add Galileo ephemeris support and mixed navigation files ([0c00209](https://github.com/JuliaGNSS/RINEXParser.jl/commit/0c002094556e9cc62447e29637f028f370408380))
+  - add Galileo ephemeris support and mixed navigation files ([0c00209](https://github.com/JuliaGNSS/RINEXParser.jl/commit/0c002094556e9cc62447e29637f028f370408380))